+++ src/app/chat/ChatInfo/PostureList/SideBar/index.tsx
@@ -1,35 +1,24 @@
 import { Divider } from 'antd';
-import { createStyles } from 'antd-style';
-import classNames from 'classnames';
-import { ReactNode } from 'react';
+import { ReactNode, memo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Flexbox } from 'react-layout-kit';
 
+import CategoryFilter from '@/app/chat/ChatInfo/PostureList/SideBar/filters/CategoryFilter';
+import GenderFilter from '@/app/chat/ChatInfo/PostureList/SideBar/filters/GenderFilter';
 import Header from '@/components/Header';
-import ListItem from '@/components/ListItem';
 import { GenderEnum } from '@/types/agent';
 import { PostureCategoryEnum } from '@/types/touch';
 
-const useStyles = createStyles(({ css, token }) => ({
-  item: css`
-    position: relative;
-    width: 100px;
-    margin-block: 2px;
-    border-radius: ${token.borderRadius}px;
-  `,
-}));
-
 interface IndexProps {
   categoryOptions: { icon: ReactNode; label: string; value: PostureCategoryEnum | undefined }[];
-  currentCategory?: PostureCategoryEnum;
-  currentGender?: GenderEnum;
+  currentCategory: PostureCategoryEnum | undefined;
+  currentGender: GenderEnum | undefined;
   genderOptions: { icon: ReactNode; label: string; value: GenderEnum | undefined }[];
   setCurrentCategory: (category?: PostureCategoryEnum) => void;
   setCurrentGender: (gender?: GenderEnum) => void;
 }
 
-const Index = (props: IndexProps) => {
-  const { styles } = useStyles();
+const Index = memo((props: IndexProps) => {
   const {
     currentGender,
     currentCategory,
@@ -43,30 +32,21 @@ const Index = (props: IndexProps) => {
   return (
     <Flexbox>
       <Header title={t('info.genderLabel')} />
-      {genderOptions.map((item) => (
-        <ListItem
-          avatar={item.icon}
-          className={classNames(styles.item)}
-          active={item.value === currentGender}
-          key={`gender-${item.value}`}
-          title={item.label}
-          onClick={() => setCurrentGender(item.value)}
-        />
-      ))}
+      <GenderFilter
+        genderOptions={genderOptions}
+        setCurrentGender={setCurrentGender}
+        currentGender={currentGender}
+      />
+
       <Divider />
       <Header title={t('info.postureCategoryLabel')} />
-      {categoryOptions.map((item) => (
-        <ListItem
-          avatar={item.icon}
-          className={classNames(styles.item)}
-          active={item.value === currentCategory}
-          key={`category-${item.value}`}
-          title={item.label}
-          onClick={() => setCurrentCategory(item.value)}
-        />
-      ))}
+      <CategoryFilter
+        categoryOptions={categoryOptions}
+        currentCategory={currentCategory}
+        setCurrentCategory={setCurrentCategory}
+      />
     </Flexbox>
   );
-};
+});
 
 export default Index;
